Fix URLs in TOC
